Refactor cross-check function
